requirements.txt
discord.py
gTTS
python-dotenv
aiohttp
requests